docs : Toolbar 모아서 정리
