+++ app/src/main/java/com/incava/gangchuplace/adapter/ToolbarBindingAdapter.kt
@@ -1,0 +1,68 @@
+package com.incava.gangchuplace.adapter
+
+import android.widget.Toast
+import androidx.appcompat.widget.Toolbar
+import androidx.databinding.BindingAdapter
+import androidx.navigation.findNavController
+import androidx.navigation.ui.setupWithNavController
+import com.incava.gangchuplace.model.ReviewInfo
+import com.incava.gangchuplace.viewmodel.GangChuViewModel
+import com.incava.gangchuplace.viewmodel.MyInfoViewModel
+import com.incava.gangchuplace.viewmodel.WriteViewModel
+
+object ToolbarBindingAdapter {
+
+    //GangChuViewModel과 gangchu_menu를 연결하는 바인딩어댑터
+    @JvmStatic
+    @BindingAdapter("setGangChuToolbar")
+    fun setGangChuToolbar (view : Toolbar, vm : GangChuViewModel){
+        view.setOnMenuItemClickListener { vm.itemClick(view,it)
+            true }
+    }
+
+    //Todo 툴바 링크 전송 관련.
+    @JvmStatic
+    @BindingAdapter("setReviewPageToolbar")
+    fun setReviewPageToolbar(view : Toolbar,reviewInfo: ReviewInfo){
+        view.apply {
+            setupWithNavController(findNavController())
+            setOnMenuItemClickListener {
+                Toast.makeText(context, "공유 클릭!", Toast.LENGTH_SHORT).show()
+                true
+            }
+        }
+    }
+
+    @JvmStatic
+    @BindingAdapter("setWriteToolbar")
+    fun setWriteToolbar(view: Toolbar, writeViewModel: WriteViewModel) {
+        view.setupWithNavController(view.findNavController())
+        view.setOnMenuItemClickListener {
+            writeViewModel.finishReview(view)
+            true
+        }
+    }
+
+    @JvmStatic
+    @BindingAdapter("setReviseInfoToolbar")
+    fun setReviseInfoToolbar(view: Toolbar, myInfoViewModel: MyInfoViewModel) {
+        view.setupWithNavController(view.findNavController())
+        view.setOnMenuItemClickListener {
+            //todo 리스너 구현
+            myInfoViewModel.saveInfo(view)
+            true
+        }
+    }
+
+    @JvmStatic
+    @BindingAdapter("setRevisePassToolbar")
+    fun setRevisePassToolbar(view: Toolbar, myInfoViewModel: MyInfoViewModel) {
+        view.setupWithNavController(view.findNavController())
+        view.setOnMenuItemClickListener {
+            //todo 리스너 구현
+            myInfoViewModel.saveInfo(view)
+            true
+        }
+    }
+
+}
