--- app/src/main/java/com/incava/gangchuplace/adapter/WriteBindingAdapter.kt
@@ -1,23 +1,0 @@
-package com.incava.gangchuplace.adapter
-
-import androidx.appcompat.widget.Toolbar
-import androidx.databinding.BindingAdapter
-import androidx.navigation.findNavController
-import androidx.navigation.ui.setupWithNavController
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
-import com.incava.gangchuplace.model.StorePlace
-import com.incava.gangchuplace.viewmodel.WriteViewModel
-
-object WriteBindingAdapter {
-    @JvmStatic
-    @BindingAdapter("setToolbar")
-    fun setToolbar(view: Toolbar, writeViewModel: WriteViewModel) {
-        view.setupWithNavController(view.findNavController())
-        view.setOnMenuItemClickListener {
-            writeViewModel.finishReview(view)
-            true
-        }
-    }
-
-}
